pyproject.toml
[project]
name = "eevee"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [{ name = "narumi", email = "toucans-cutouts0f@icloud.com" }]
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "feedparser>=6.0.11",
    "httpx>=0.28.0",
    "lazyopenai>=0.2.0",
    "loguru>=0.7.2",
    "pydantic>=2.10.2",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
]

[project.scripts]
eevee = "eevee.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["mypy>=1.13.0", "pytest>=8.3.3", "pytest-cov>=6.0.0", "ruff>=0.7.3"]

[tool.ruff]
exclude = ["build"]
line-length = 120

[tool.ruff.lint]
select = [
    "B",   # flake8-bugbear
    "C",   # flake8-comprehensions
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "SIM", # flake8-simplify
    "UP",  # pyupgrade
    "W",   # pycodestyle warnings
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
ignore_missing_imports = true
